backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
mistralai
numpy
scikit-learn
python-dotenv
